move logic out of the UI into the state class
